a lot more fixes...

+++ src/array/cpu/rowwise_etype_sampling_int16.cc
@@ -10,50 +10,49 @@ namespace aten {
 namespace impl {
 
 // (BarclayII) Ideally we want to have a code generator that generates these template
-// instantiations with all the combination of IdxType, FloatType and EType.  This
-// is not necessary for the current milestone (2022.9.30) though.
-template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int32_t, float, int32_t>(
+// instantiations with all the combination of IdxType, FloatType and EType.
+template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int32_t, float, int16_t>(
     CSRMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int64_t, float, int64_t>(
+template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int64_t, float, int16_t>(
     CSRMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int32_t, double, int32_t>(
+template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int32_t, double, int16_t>(
     CSRMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int64_t, double, int64_t>(
+template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int64_t, double, int16_t>(
     CSRMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int32_t, int8_t, int32_t>(
+template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int32_t, int8_t, int16_t>(
     CSRMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int64_t, int8_t, int64_t>(
+template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int64_t, int8_t, int16_t>(
     CSRMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int32_t, uint8_t, int32_t>(
+template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int32_t, uint8_t, int16_t>(
     CSRMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int64_t, uint8_t, int64_t>(
+template COOMatrix CSRRowWisePerEtypeSampling<kDGLCPU, int64_t, uint8_t, int16_t>(
     CSRMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
 
-template COOMatrix CSRRowWisePerEtypeSamplingUniform<kDGLCPU, int32_t, int32_t>(
+template COOMatrix CSRRowWisePerEtypeSamplingUniform<kDGLCPU, int32_t, int16_t>(
     CSRMatrix, IdArray, IdArray, const std::vector<int64_t>&, bool, bool);
-template COOMatrix CSRRowWisePerEtypeSamplingUniform<kDGLCPU, int64_t, int64_t>(
+template COOMatrix CSRRowWisePerEtypeSamplingUniform<kDGLCPU, int64_t, int16_t>(
     CSRMatrix, IdArray, IdArray, const std::vector<int64_t>&, bool, bool);
 
-template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int32_t, float, int32_t>(
+template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int32_t, float, int16_t>(
     COOMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int64_t, float, int64_t>(
+template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int64_t, float, int16_t>(
     COOMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int32_t, double, int32_t>(
+template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int32_t, double, int16_t>(
     COOMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int64_t, double, int64_t>(
+template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int64_t, double, int16_t>(
     COOMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int32_t, int8_t, int32_t>(
+template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int32_t, int8_t, int16_t>(
     COOMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int64_t, int8_t, int64_t>(
+template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int64_t, int8_t, int16_t>(
     COOMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int32_t, uint8_t, int32_t>(
+template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int32_t, uint8_t, int16_t>(
     COOMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
-template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int64_t, uint8_t, int64_t>(
+template COOMatrix COORowWisePerEtypeSampling<kDGLCPU, int64_t, uint8_t, int16_t>(
     COOMatrix, IdArray, IdArray, const std::vector<int64_t>&, FloatArray, bool, bool);
 
-template COOMatrix COORowWisePerEtypeSamplingUniform<kDGLCPU, int32_t, int32_t>(
+template COOMatrix COORowWisePerEtypeSamplingUniform<kDGLCPU, int32_t, int16_t>(
     COOMatrix, IdArray, IdArray, const std::vector<int64_t>&, bool, bool);
-template COOMatrix COORowWisePerEtypeSamplingUniform<kDGLCPU, int64_t, int64_t>(
+template COOMatrix COORowWisePerEtypeSamplingUniform<kDGLCPU, int64_t, int16_t>(
     COOMatrix, IdArray, IdArray, const std::vector<int64_t>&, bool, bool);
 }  // namespace impl
 }  // namespace aten
